Fix #75: Refactor predictoor approach directory structure
